requirements.txt
pydantic>=2.0.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
ortools
google-generativeai
pytest
